+++ lib/omniauth-shippo.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require 'omniauth-shippo/version'
+require 'omniauth/strategies/shippo'