+++ lib/omniauth-tradegecko.rb
@@ -1,2 +1,0 @@
-require "omniauth-tradegecko/version"
-require 'omniauth/strategies/tradegecko'