+++ spec/spec_helper.rb
@@ -6,7 +6,7 @@ require 'rspec'
 require 'rack/test'
 require 'webmock/rspec'
 require 'omniauth'
-require 'omniauth-tradegecko'
+require 'omniauth-shippo'
 
 RSpec.configure do |config|
   config.include WebMock::API