+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-## 1.0.0 (2018-04-11)
-- Unlock omniauth-oauth 1.3.1+ by hardcoding the exclusion of query-params
+## Unreleased
+- Initial release - repurpose omniauth-tradegecko